refactor(node): remove obsolete benchmark paths
