+++ crates/node/README.md
@@ -17,6 +17,10 @@ through the `ZmqPublisher` trait and its `SocketZmqPublisher` / `TracingZmqPubli
 / `NoOpZmqPublisher` implementations and the `TxIndexRuntime` worker; `signal` and
 `shutdown` bridge process signals into graceful shutdown.
 
+The node crate registers only `benches/sync_pipeline.rs` as a Criterion benchmark.
+Corpus export and mainnet replay measurement live under `examples/`, so diagnostic
+recorders and RSS sampling cannot become part of the runtime crate API.
+
 ## Features
 - `default` (enables `fjall` and `kernel`): the performance-oriented fjall storage
   backend plus the bitcoinkernel consensus verifier, so per-crate `cargo check` works
@@ -31,7 +35,7 @@ through the `ZmqPublisher` trait and its `SocketZmqPublisher` / `TracingZmqPubli
 - `mimalloc`: pulls the optional `mimalloc` dependency; the
   `mainnet_prefix_replay` example registers it as the global allocator.
 - `prometheus-http`: enables the `metrics-exporter-prometheus/http-listener` feature;
-  the in-process metrics recorder does not start an HTTP listener.
+  the production listener itself is controlled by `metrics_bind`.
 
 Part of [`bitcoin-rs`](../../README.md); see [`CONCEPTS.md`](../../CONCEPTS.md) for the
 project vocabulary.
