+++ crates/utxo/benches/utxo_commit.rs
@@ -10,16 +10,11 @@
 #[global_allocator]
 static GLOBAL_MIMALLOC: mimalloc::MiMalloc = mimalloc::MiMalloc;
 
-use std::{
-    hint::black_box,
-    time::{Duration, Instant},
-};
+use std::hint::black_box;
 
 use bitcoin::{Amount, ScriptBuf};
 use bitcoin_rs_primitives::{Hash256, OutPoint, TxOut};
-use bitcoin_rs_utxo::{
-    BlockChanges, UtxoAdd, UtxoChangeEvents, UtxoChangeListener, UtxoInserted, UtxoRemoved, UtxoSet,
-};
+use bitcoin_rs_utxo::{BlockChanges, UtxoAdd, UtxoSet};
 use criterion::{BatchSize, Criterion, criterion_group, criterion_main};
 
 const ENTRY_COUNT: u64 = 10_000;
@@ -34,7 +29,6 @@ const SPEND_PROXY_SPEND_OUTPUT_VALUE: u64 = 78_124_999;
 #[derive(Copy, Clone, Debug)]
 enum ShardShape {
     Existing,
-    Uniform,
     TwoShard,
     FourShard,
     Concentrated,
@@ -51,17 +45,6 @@ struct SyntheticEntry {
 struct SyntheticWorkload {
     spends: Vec<SyntheticEntry>,
     adds: Vec<SyntheticEntry>,
-    distribution: [usize; 256],
-}
-
-struct NoopListener;
-
-impl UtxoChangeListener for NoopListener {
-    fn on_insert_coins(&self, _insertions: &[UtxoInserted<'_>]) {}
-
-    fn on_remove_coins(&self, _removals: &[UtxoRemoved]) {}
-
-    fn on_committed_event_batches(&self, _batches: &[UtxoChangeEvents<'_>]) {}
 }
 
 const fn next_u64(state: &mut u64) -> u64 {
@@ -84,11 +67,6 @@ fn shaped_txid(seed: u64, index: u64, shape: ShardShape) -> Hash256 {
     let mut hash = txid(seed);
     match shape {
         ShardShape::Existing => {}
-        ShardShape::Uniform => {
-            let mut bytes = hash.to_le_bytes();
-            bytes[0] = u8::try_from(index % 256).unwrap_or(0);
-            hash = Hash256::from_le_bytes(&bytes);
-        }
         ShardShape::TwoShard => {
             let mut bytes = hash.to_le_bytes();
             bytes[0] = u8::try_from(index % 2).unwrap_or(0);
@@ -122,7 +100,6 @@ fn synthetic_workload(seed: u64, shape: ShardShape) -> SyntheticWorkload {
     let mut rng = seed;
     let mut spends = Vec::with_capacity(usize::try_from(ENTRY_COUNT).unwrap_or(0));
     let mut adds = Vec::with_capacity(usize::try_from(ENTRY_COUNT).unwrap_or(0));
-    let mut distribution = [0_usize; 256];
 
     for i in 0_u64..ENTRY_COUNT {
         let spend_seed = next_u64(&mut rng);
@@ -138,8 +115,6 @@ fn synthetic_workload(seed: u64, shape: ShardShape) -> SyntheticWorkload {
     for i in 0_u64..ENTRY_COUNT {
         let add_seed = next_u64(&mut rng).wrapping_add(i);
         let outpoint = OutPoint::new(shaped_txid(add_seed, i, shape), 0);
-        let shard = usize::from(outpoint.txid.prefix8()[0]);
-        distribution[shard] = distribution[shard].saturating_add(1);
         adds.push(SyntheticEntry {
             outpoint,
             txout: txout(add_seed),
@@ -148,11 +123,7 @@ fn synthetic_workload(seed: u64, shape: ShardShape) -> SyntheticWorkload {
         });
     }
 
-    SyntheticWorkload {
-        spends,
-        adds,
-        distribution,
-    }
+    SyntheticWorkload { spends, adds }
 }
 
 fn preload_set(workload: &SyntheticWorkload, seed: u64) -> UtxoSet {
@@ -358,17 +329,9 @@ fn utxo_add(entry: &SyntheticEntry) -> UtxoAdd {
     )
 }
 
-fn synthetic_case(seed: u64, shape: ShardShape) -> (UtxoSet, BlockChanges, [usize; 256]) {
+fn synthetic_case(seed: u64, shape: ShardShape) -> (UtxoSet, BlockChanges) {
     let workload = synthetic_workload(seed, shape);
     let set = preload_set(&workload, seed);
-    let changes = block_changes(&workload);
-    (set, changes, workload.distribution)
-}
-
-fn synthetic_listener_case(seed: u64, shape: ShardShape) -> (UtxoSet, BlockChanges) {
-    let workload = synthetic_workload(seed, shape);
-    let mut set = preload_set(&workload, seed);
-    set.set_listener(Box::new(NoopListener));
     let changes = block_changes(&workload);
     (set, changes)
 }
@@ -387,97 +350,11 @@ fn spend_proxy_spend_txout() -> TxOut {
     }
 }
 
-fn summarize_distribution(distribution: &[usize; 256]) -> (usize, usize, usize) {
-    let mut active = 0_usize;
-    let mut min_non_zero = usize::MAX;
-    let mut max = 0_usize;
-    for &count in distribution {
-        if count == 0 {
-            continue;
-        }
-        active = active.saturating_add(1);
-        min_non_zero = min_non_zero.min(count);
-        max = max.max(count);
-    }
-    if active == 0 {
-        min_non_zero = 0;
-    }
-    (active, min_non_zero, max)
-}
-
-fn distribution_prepass(seed: u64, shape: ShardShape) -> [usize; 256] {
-    let mut rng = seed;
-    for _ in 0_u64..ENTRY_COUNT {
-        let _ = next_u64(&mut rng);
-    }
-    let mut distribution = [0_usize; 256];
-    for i in 0_u64..ENTRY_COUNT {
-        let add_seed = next_u64(&mut rng).wrapping_add(i);
-        let txid = shaped_txid(add_seed, i, shape);
-        let shard = usize::from(txid.prefix8()[0]);
-        distribution[shard] = distribution[shard].saturating_add(1);
-    }
-    distribution
-}
-
-const fn percentile(samples: &[Duration], numerator: usize, denominator: usize) -> Duration {
-    if samples.is_empty() {
-        return Duration::ZERO;
-    }
-    let last = samples.len().saturating_sub(1);
-    let index = last.saturating_mul(numerator) / denominator;
-    samples[index]
-}
-
-fn print_synthetic_summary(name: &str, shape: ShardShape) {
-    let mut prepass_samples = Vec::with_capacity(9);
-    let mut commit_samples = Vec::with_capacity(9);
-    let distribution = distribution_prepass(0x5555_aaaa_ffff_0000, shape);
-    for seed in 0_u64..9 {
-        let prepass_start = Instant::now();
-        let _ = distribution_prepass(seed + 1, shape);
-        prepass_samples.push(prepass_start.elapsed());
-
-        let (set, changes, _) = synthetic_case(seed + 1, shape);
-        let start = Instant::now();
-        if let Err(error) = set.commit_block(black_box(&changes), &txid(seed)) {
-            panic!("synthetic commit failed: {error}");
-        }
-        commit_samples.push(start.elapsed());
-    }
-    prepass_samples.sort_unstable();
-    commit_samples.sort_unstable();
-    let (active, min_non_zero, max) = summarize_distribution(&distribution);
-    println!(
-        "utxo_commit_{name} prepass p50={:?} p95={:?} p99={:?} commit p50={:?} p95={:?} p99={:?} active_shards={active} min_non_zero_shard_entries={min_non_zero} max_shard_entries={max}",
-        percentile(&prepass_samples, 50, 100),
-        percentile(&prepass_samples, 95, 100),
-        percentile(&prepass_samples, 99, 100),
-        percentile(&commit_samples, 50, 100),
-        percentile(&commit_samples, 95, 100),
-        percentile(&commit_samples, 99, 100),
-    );
-}
-
-fn bench_uniform_noop_listener(c: &mut Criterion) {
-    c.bench_function("utxo_commit/uniform_noop_listener", |b| {
-        b.iter_batched(
-            || synthetic_listener_case(0x00ab_cdef, ShardShape::Uniform),
-            |(set, changes)| {
-                if let Err(error) = set.commit_block(black_box(&changes), &txid(0x0012_3456)) {
-                    panic!("synthetic listener commit failed: {error}");
-                }
-            },
-            BatchSize::SmallInput,
-        );
-    });
-}
-
 fn bench_two_shard(c: &mut Criterion) {
     c.bench_function("utxo_commit/two_shard", |b| {
         b.iter_batched(
             || synthetic_case(0x00ab_cdef, ShardShape::TwoShard),
-            |(set, changes, _distribution)| {
+            |(set, changes)| {
                 if let Err(error) = set.commit_block(black_box(&changes), &txid(0x0012_3456)) {
                     panic!("synthetic two-shard commit failed: {error}");
                 }
@@ -491,55 +368,9 @@ fn bench_four_shard(c: &mut Criterion) {
     c.bench_function("utxo_commit/four_shard", |b| {
         b.iter_batched(
             || synthetic_case(0x00ab_cdef, ShardShape::FourShard),
-            |(set, changes, _distribution)| {
+            |(set, changes)| {
                 if let Err(error) = set.commit_block(black_box(&changes), &txid(0x0012_3456)) {
                     panic!("synthetic four-shard commit failed: {error}");
-                }
-            },
-            BatchSize::SmallInput,
-        );
-    });
-}
-
-fn bench_two_shard_noop_listener(c: &mut Criterion) {
-    c.bench_function("utxo_commit/two_shard_noop_listener", |b| {
-        b.iter_batched(
-            || synthetic_listener_case(0x00ab_cdef, ShardShape::TwoShard),
-            |(set, changes)| {
-                if let Err(error) = set.commit_block(black_box(&changes), &txid(0x0012_3456)) {
-                    panic!("synthetic two-shard listener commit failed: {error}");
-                }
-            },
-            BatchSize::SmallInput,
-        );
-    });
-}
-
-fn bench_concentrated_noop_listener(c: &mut Criterion) {
-    c.bench_function("utxo_commit/concentrated_noop_listener", |b| {
-        b.iter_batched(
-            || synthetic_listener_case(0x00ab_cdef, ShardShape::Concentrated),
-            |(set, changes)| {
-                if let Err(error) = set.commit_block(black_box(&changes), &txid(0x0012_3456)) {
-                    panic!("synthetic concentrated listener commit failed: {error}");
-                }
-            },
-            BatchSize::SmallInput,
-        );
-    });
-}
-
-fn bench_same_txid_churn_noop_listener(c: &mut Criterion) {
-    c.bench_function("utxo_commit/same_txid_churn_noop_listener", |b| {
-        b.iter_batched(
-            || {
-                let (mut set, changes) = same_txid_churn_case(0x0102_0304);
-                set.set_listener(Box::new(NoopListener));
-                (set, changes)
-            },
-            |(set, changes)| {
-                if let Err(error) = set.commit_block(black_box(&changes), &txid(0x0112_1314)) {
-                    panic!("same-txid churn listener commit failed: {error}");
                 }
             },
             BatchSize::SmallInput,
@@ -561,24 +392,6 @@ fn bench_same_txid_full_spend(c: &mut Criterion) {
     });
 }
 
-fn bench_same_txid_full_spend_noop_listener(c: &mut Criterion) {
-    c.bench_function("utxo_commit/same_txid_full_spend_noop_listener", |b| {
-        b.iter_batched(
-            || {
-                let (mut set, changes) = same_txid_full_spend_case(0x0203_0405);
-                set.set_listener(Box::new(NoopListener));
-                (set, changes)
-            },
-            |(set, changes)| {
-                if let Err(error) = set.commit_block(black_box(&changes), &txid(0x0212_1314)) {
-                    panic!("same-txid full-spend listener commit failed: {error}");
-                }
-            },
-            BatchSize::SmallInput,
-        );
-    });
-}
-
 fn bench_same_txid_high_vout_full_spend(c: &mut Criterion) {
     c.bench_function("utxo_commit/same_txid_high_vout_full_spend", |b| {
         b.iter_batched(
@@ -593,27 +406,6 @@ fn bench_same_txid_high_vout_full_spend(c: &mut Criterion) {
     });
 }
 
-fn bench_same_txid_high_vout_full_spend_noop_listener(c: &mut Criterion) {
-    c.bench_function(
-        "utxo_commit/same_txid_high_vout_full_spend_noop_listener",
-        |b| {
-            b.iter_batched(
-                || {
-                    let (mut set, changes) = same_txid_high_vout_full_spend_case(0x0506_0708);
-                    set.set_listener(Box::new(NoopListener));
-                    (set, changes)
-                },
-                |(set, changes)| {
-                    if let Err(error) = set.commit_block(black_box(&changes), &txid(0x0512_1314)) {
-                        panic!("same-txid high-vout full-spend listener commit failed: {error}");
-                    }
-                },
-                BatchSize::SmallInput,
-            );
-        },
-    );
-}
-
 fn bench_spend_fanout(c: &mut Criterion) {
     c.bench_function("utxo_commit/spend_fanout_64", |b| {
         b.iter_batched(
@@ -621,21 +413,6 @@ fn bench_spend_fanout(c: &mut Criterion) {
             |(set, changes)| {
                 if let Err(error) = set.commit_block(black_box(&changes), &txid(0x0412_1314)) {
                     panic!("spend-fanout commit failed: {error}");
-                }
-            },
-            BatchSize::SmallInput,
-        );
-    });
-    c.bench_function("utxo_commit/spend_fanout_64_noop_listener", |b| {
-        b.iter_batched(
-            || {
-                let (mut set, changes) = spend_fanout_case(0x0405_0607);
-                set.set_listener(Box::new(NoopListener));
-                (set, changes)
-            },
-            |(set, changes)| {
-                if let Err(error) = set.commit_block(black_box(&changes), &txid(0x0412_1314)) {
-                    panic!("spend-fanout listener commit failed: {error}");
                 }
             },
             BatchSize::SmallInput,
@@ -719,11 +496,8 @@ fn bench_same_txid_cases(c: &mut Criterion) {
             BatchSize::SmallInput,
         );
     });
-    bench_same_txid_churn_noop_listener(c);
     bench_same_txid_full_spend(c);
-    bench_same_txid_full_spend_noop_listener(c);
     bench_same_txid_high_vout_full_spend(c);
-    bench_same_txid_high_vout_full_spend_noop_listener(c);
     bench_spend_fanout(c);
     bench_matched_lookup(c);
     c.bench_function("utxo_commit/interleaved_same_txid_churn", |b| {
@@ -737,114 +511,29 @@ fn bench_same_txid_cases(c: &mut Criterion) {
             BatchSize::SmallInput,
         );
     });
-    c.bench_function(
-        "utxo_commit/interleaved_same_txid_churn_noop_listener",
-        |b| {
-            b.iter_batched(
-                || {
-                    let (mut set, changes) = interleaved_same_txid_churn_case(0x0304_0506);
-                    set.set_listener(Box::new(NoopListener));
-                    (set, changes)
-                },
-                |(set, changes)| {
-                    if let Err(error) = set.commit_block(black_box(&changes), &txid(0x0312_1314)) {
-                        panic!("interleaved same-txid churn listener commit failed: {error}");
-                    }
-                },
-                BatchSize::SmallInput,
-            );
-        },
-    );
 }
 
 fn utxo_commit_synthetic_block(c: &mut Criterion) {
-    print_synthetic_summary("existing", ShardShape::Existing);
     c.bench_function("utxo_commit/existing", |b| {
         b.iter_batched(
             || synthetic_case(0x00ab_cdef, ShardShape::Existing),
-            |(set, changes, _distribution)| {
+            |(set, changes)| {
                 if let Err(error) = set.commit_block(black_box(&changes), &txid(0x0012_3456)) {
                     panic!("synthetic commit failed: {error}");
-                }
-            },
-            BatchSize::SmallInput,
-        );
-    });
-    c.bench_function("utxo_build_commit/existing", |b| {
-        b.iter_batched(
-            || {
-                let workload = synthetic_workload(0x00ab_cdef, ShardShape::Existing);
-                let set = preload_set(&workload, 0x00ab_cdef);
-                (set, workload)
-            },
-            |(set, workload)| {
-                let changes = block_changes(black_box(&workload));
-                if let Err(error) = set.commit_block(black_box(&changes), &txid(0x0012_3456)) {
-                    panic!("synthetic build+commit failed: {error}");
                 }
             },
             BatchSize::SmallInput,
         );
     });
     bench_same_txid_cases(c);
-
-    print_synthetic_summary("uniform", ShardShape::Uniform);
-    c.bench_function("utxo_commit/uniform", |b| {
-        b.iter_batched(
-            || synthetic_case(0x00ab_cdef, ShardShape::Uniform),
-            |(set, changes, _distribution)| {
-                if let Err(error) = set.commit_block(black_box(&changes), &txid(0x0012_3456)) {
-                    panic!("synthetic commit failed: {error}");
-                }
-            },
-            BatchSize::SmallInput,
-        );
-    });
-    bench_uniform_noop_listener(c);
     bench_two_shard(c);
     bench_four_shard(c);
-    bench_two_shard_noop_listener(c);
-    c.bench_function("utxo_build_commit/uniform", |b| {
-        b.iter_batched(
-            || {
-                let workload = synthetic_workload(0x00ab_cdef, ShardShape::Uniform);
-                let set = preload_set(&workload, 0x00ab_cdef);
-                (set, workload)
-            },
-            |(set, workload)| {
-                let changes = block_changes(black_box(&workload));
-                if let Err(error) = set.commit_block(black_box(&changes), &txid(0x0012_3456)) {
-                    panic!("synthetic build+commit failed: {error}");
-                }
-            },
-            BatchSize::SmallInput,
-        );
-    });
-
-    print_synthetic_summary("concentrated", ShardShape::Concentrated);
     c.bench_function("utxo_commit/concentrated", |b| {
         b.iter_batched(
             || synthetic_case(0x00ab_cdef, ShardShape::Concentrated),
-            |(set, changes, _distribution)| {
+            |(set, changes)| {
                 if let Err(error) = set.commit_block(black_box(&changes), &txid(0x0012_3456)) {
                     panic!("synthetic commit failed: {error}");
-                }
-            },
-            BatchSize::SmallInput,
-        );
-    });
-    bench_concentrated_noop_listener(c);
-    c.bench_function("utxo_build_commit/concentrated", |b| {
-        b.iter_batched(
-            || {
-                let workload = synthetic_workload(0x00ab_cdef, ShardShape::Concentrated);
-                let set = preload_set(&workload, 0x00ab_cdef);
-                (set, workload)
-            },
-            |(set, workload)| {
-                let changes = block_changes(black_box(&workload));
-                if let Err(error) = set.commit_block(black_box(&changes), &txid(0x0012_3456)) {
-                    panic!("synthetic build+commit failed: {error}");
                 }
             },
             BatchSize::SmallInput,
