+++ docs/solutions/architecture-patterns/node-reorg-execution-design.md
@@ -176,7 +176,7 @@ Done:
 | Persistence | queued before the block body and UTXO commit; flushed with a clean checkpoint, not per block |
 | `disconnect_block` | restores the UTXO set, coinstats, and applied tip; then publishes a TxIndex wake |
 | TxIndex reconciliation | one node-owned worker rolls enabled capability watermarks back to the common ancestor, then assembles count-and-byte-bounded forward batches across strict descendant tip revisions. Equal cursors share one body parse and atomic commit; divergent cursors move independently. `--scriptindex` enables generic script UTXO, history, and outpoint-spender queries, while only explicit `--txindex` advertises Core txindex semantics. Each queued wake triggers reconciliation without moving the pending batch's fixed deadline. Exact capability snapshot-plus-tip gating keeps individual queries unavailable until every consumed watermark matches. Esplora additionally captures the applied-tip identity around each GET response so independently safe index reads cannot be combined across a reorg. |
-| `coin_stats` rewind | block-level fields only; the per-coin ones ride the `UtxoSet` change listener, which the undo already drives in reverse |
+| `coin_stats` rewind | block-level fields are explicitly rewound; optional listener users get per-coin inverse updates from the UTXO undo, while the default node recomputes UTXO-derived fields during checkpoint and stable reads |
 | Filter header cache | repointed at the parent; the index itself needs no rollback because its rows are hash-addressed like block bodies |
 | `blocks` RPC cache | popped when the tail is ours; absence is legitimate after a restart or a prune |
 | `DisconnectError` | splits `Refused` (nothing touched) from `Fatal` (partly rolled back, carries hash and height), plus `MarkerStuck` (rolled back, but the interlock would not clear) |
